Log when postMessageDirect browser is null
